move project parsing events (I) to core
