+++ core/dbt/config/runtime.py
@@ -24,7 +24,7 @@ from dbt.contracts.graph.manifest import ManifestMetadata
 from dbt.contracts.project import Configuration, UserConfig
 from dbt.common.dataclass_schema import ValidationError
 from dbt.common.events.functions import warn_or_error
-from dbt.common.events.types import UnusedResourceConfigPath
+from dbt.events.types import UnusedResourceConfigPath
 from dbt.exceptions import (
     ConfigContractBrokenError,
     DbtProjectError,
