+++ core/dbt/contracts/graph/manifest.py
@@ -63,9 +63,8 @@ from dbt.exceptions import (
 from dbt.adapters.exceptions import DuplicateMacroInPackageError, DuplicateMaterializationNameError
 from dbt.common.helper_types import PathSet
 from dbt.common.events.functions import fire_event
-from dbt.common.events.types import UnpinnedRefNewVersionAvailable
 from dbt.common.events.contextvars import get_node_info
-from dbt.events.types import MergedFromState
+from dbt.events.types import MergedFromState, UnpinnedRefNewVersionAvailable
 from dbt.node_types import NodeType, AccessType
 from dbt.flags import get_flags
 from dbt.mp_context import get_mp_context
