+++ core/dbt/contracts/graph/nodes.py
@@ -47,7 +47,7 @@ from dbt.contracts.util import Replaceable
 from dbt.common.contracts.config.properties import AdditionalPropertiesMixin
 from dbt.common.events.functions import warn_or_error
 from dbt.exceptions import ParsingError, ContractBreakingChangeError
-from dbt.common.events.types import (
+from dbt.events.types import (
     SeedIncreased,
     SeedExceedsLimitSamePath,
     SeedExceedsLimitAndPathChanged,
