+++ core/dbt/contracts/graph/semantic_manifest.py
@@ -16,7 +16,7 @@ from dbt_semantic_interfaces.validations.semantic_manifest_validator import (
 from dbt.common.clients.system import write_file
 from dbt.common.events.base_types import EventLevel
 from dbt.common.events.functions import fire_event
-from dbt.common.events.types import SemanticValidationFailure
+from dbt.events.types import SemanticValidationFailure
 from dbt.exceptions import ParsingError
