+++ core/dbt/contracts/state.py
@@ -5,7 +5,7 @@ from dbt.contracts.graph.manifest import WritableManifest
 from dbt.contracts.results import FreshnessExecutionResultArtifact
 from dbt.contracts.results import RunResultsArtifact
 from dbt.common.events.functions import fire_event
-from dbt.common.events.types import WarnStateTargetEqual
+from dbt.events.types import WarnStateTargetEqual
 from dbt.exceptions import IncompatibleSchemaError
