+++ core/dbt/parser/manifest.py
@@ -44,6 +44,9 @@ from dbt.constants import (
 from dbt.common.helper_types import PathSet
 from dbt.common.events.functions import fire_event, get_invocation_id, warn_or_error
 from dbt.common.events.types import (
+    Note,
+)
+from dbt.events.types import (
     PartialParsingErrorProcessingFile,
     PartialParsingError,
     ParsePerfInfoPath,
@@ -54,12 +57,9 @@ from dbt.common.events.types import (
     InvalidDisabledTargetInTestNode,
     NodeNotFoundOrDisabled,
     StateCheckVarsHash,
-    Note,
+    DeprecatedModel,
     DeprecatedReference,
     UpcomingReferenceDeprecation,
-)
-from dbt.events.types import (
-    DeprecatedModel,
 )
 from dbt.logger import DbtProcessState
 from dbt.node_types import NodeType, AccessType
