+++ core/dbt/parser/partial.py
@@ -10,7 +10,7 @@ from dbt.contracts.files import (
 )
 from dbt.common.events.functions import fire_event
 from dbt.common.events.base_types import EventLevel
-from dbt.common.events.types import (
+from dbt.events.types import (
     PartialParsingEnabled,
     PartialParsingFile,
 )
